Flask-Nft/pyproject.toml
[tool.poetry]
name = "flask-nft"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-restful = "^0.3.10"
pillow = "^11.1.0"
flask-uploads = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
